Rewrite python binding CMake configuration and CI checks
This does three things:

1. Complete rewrite of the CMake configuration for the python
   bindings. This drops support for Python2, and it eliminates
   src/python/config/ModuleDefine.cmake altogether and moves the CMake
   code directly into PyImath, PyImathNumpy, etc. It's more direct and
   easier to follow this way.

2. Add CI checks to build and test the python bindings on macOS and
   Windows, previously not done. The python bindings require
   boost.python, which is now installed via conan. A few caveats on
   Windows:

   a. The Windows Debug build does not currently work. It stubbornly
      requires the debug library python39_d.lib, and I can't figure
      out how to redirect it to python39.lib. None of the suggested
      strategies seem to work, so the build is disabled for now. This
      is left as a TODO.

   b. The python and pybind11 bindings build successfully on Windows,
      but they don't currently execute properly, the "import" checks
      fail. I suspect a link/symbol error, but the message is unhelpful. The
      PATH and PYTHONPATH appear to be set correctly and as far as I
      can tell the necessary dlls are in place. This is left as a TODO
      (several TODO's actually).

3. Configure PyImath as a COMPONENT of Imath, to address #360 and
   #395. Downstream projects that require PyImath must now specify:

      find_package(Imath COMPONENTS PyImath)

   whereas projects that simply say:

      find_package(Imath)

   will no longer generate a dependency on PyImath. Previously, the
   only way to avoid this dependency was to build/install Imath with
   the python bindings disabled.  #360 was based on the hope/assumption
   that optional downstream dependence on the python bindings could be
   configured at Imath build/install time, but the desired behavior
   can only be accomplished by introducing a new downstream
   configuration option.

   The CI validates this by building the ImathTest and PyImathTest
   projects as standalone applications linked against the
   just-installed Imath/PyImath build, confirming both that they
   compile/link and that they have the proper dependencies.

+++ src/pybind11/PyBindImath/PyBindImathVec.cpp
@@ -156,12 +156,12 @@ register_vec(py::class_<Vec>& c)
         .def("__getitem__", [](const Vec &v, size_t i) {
             if (i < 0 || i >= v.dimensions())
                 throw std::domain_error ("invalid index");
-            return v[i];
+            return v[static_cast<int>(i)];
         })
         .def("__setitem__", [](Vec &v, size_t i, T value) {
             if (i < 0 || i >= v.dimensions())
                 throw std::domain_error ("invalid index");
-            v[i] = value;
+            v[static_cast<int>(i)] = value;
         })
         .def(py::self != py::self)
         .def(py::self ^ py::self)
